requirements.txt
pyfirmata
pysine
pyserial
